+++ backend/src/main/java/wooteco/retrospective/presentation/dto/attendance/AttendanceRequest.java
@@ -1,4 +1,4 @@
-package wooteco.retrospective.dto;
+package wooteco.retrospective.presentation.dto.attendance;
 
 public class AttendanceRequest {
